source/cloudfunction_main/requirements.txt
google-cloud-pubsub
google-cloud-datastore
oauth2client
pydrive